Refactor out the binary formatter.
