+++ ExamplePythonDependency/quick_demo.py
@@ -3,3 +3,4 @@ def scream(letter: str, length: int) -> str:
 
 def scream_names(names: list[str], length: int) -> list[str]:
     return [scream(name, length) for name in names]
+
