+++ PythonEnvironments/PythonEnvironments.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="pythonnet" Version="3.0.3" />
+    <PackageReference Include="pythonnet" Version="3.1.0-preview2024-06-03" />
   </ItemGroup>
 
 </Project>
