+++ PythonSourceGenerator/PythonSourceGenerator.csproj
@@ -13,7 +13,7 @@
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
-        <PackageReference Include="pythonnet" Version="3.0.3">
+        <PackageReference Include="pythonnet" Version="3.1.0-preview2024-06-03">
             <GeneratePathProperty>true</GeneratePathProperty>
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive; compile</IncludeAssets>
